split conftest.py to avoid some cyclic import
